requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
pydantic>=2.0
tqdm>=4.65
aiosqlite>=0.19.0
